# docs/api-specs.yml
openapi: "3.0.2"
info:
  title: Tabungan API
  version: "1.0"
servers:
  - url: https://api.server.test/v1
paths:
  /daftar:
    post:
      tags:
        - "Tabungan API"
      description: Register new Account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama:
                  type: string
                nik:
                  type: string
                no_hp:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      no_rekening:
                        type: string
        "400":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  remark:
                    type: object
                    properties:
                      err_message:
                        type: string
        
  /tabung:
    post:
      tags:
        - "Tabungan API"
      description: TopUp Cash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nominal:
                  type: number
                no_rekening:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      saldo:
                        type: number
        "400":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  remark:
                    type: object
                    properties:
                      err_message:
                        type: string
        
  /tarik:
    post:
      tags:
        - "Tabungan API"
      description: Withdrawal Cash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nominal:
                  type: number
                no_rekening:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      saldo:
                        type: number
        "400":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  remark:
                    type: object
                    properties:
                      err_message:
                        type: string
  
  /saldo/<no_rekening>:
    get:
      tags:
        - "Tabungan API"
      description: Check Balance
      parameters:
        - name: no_rekening
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      saldo:
                        type: number
        "400":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  remark:
                    type: object
                    properties:
                      err_message:
                        type: string
        
  /mutasi/<no_rekening>:
    get:
      tags:
        - "Tabungan API"
      description: Check Mutations
      parameters:
        - name: no_rekening
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        kode_transaksi:
                          type: string
                        nominal:
                          type: number
                        waktu:
                          type: string
        "400":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  remark:
                    type: object
                    properties:
                      err_message:
                        type: string
        